allow socks5h as an alternate scheme for socks5
scoks5h is a convention in the curl world to indicate "use" of
proxy-host's dns to resolve the address of the destination address.

Since python-socks already has implemented support for
[DOMAIN](https://github.com/romis2012/python-socks/blob/master/python_socks/_proto/socks5.py#L20)
this additional scheme type can also be allowed.

this allows not having to maintain 2 url (with and without socck5h) in
the users of aiohttps-socks/python-socks/.. and requests